# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing the 3X-UI panel including inbounds, server, and settings.",
        "title": "3X-UI Panel API",
        "contact": {
            "name": "3X-UI"
        },
        "license": {
            "name": "GPL-3.0"
        },
        "version": "2.0"
    },
    "basePath": "/",
    "paths": {
...
        "/panel/setting/apikeys/create": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ApiKey"
                ],
                "summary": "Create a new API key",
                "parameters": [
                    {
                        "description": "API key details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.createApiKeyForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Msg"
                        }
                    }
                }
            }
        },
        "/panel/setting/apikeys/delete/{id}": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ApiKey"
                ],
                "summary": "Delete an API key",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "API Key ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Msg"
                        }
                    }
                }
            }
        },
...
        "/panel/setting/apikeys/toggle/{id}": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ApiKey"
                ],
                "summary": "Toggle an API key enabled/disabled",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "API Key ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Msg"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controller.createApiKeyForm": {
            "type": "object",
            "required": [
                "description"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "expiryTime": {
                    "type": "integer"
                }
            }
        },
        "entity.Msg": {
            "type": "object",
            "properties": {
                "msg": {
                    "description": "Response message text",
                    "type": "string"
                },
                "obj": {
                    "description": "Optional data object"
                },
                "success": {
                    "description": "Indicates if the operation was successful",
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}